+++ morpheus-graphql-server/src/Data/Morpheus/Server/Types/DirectiveDefinitions.hs
@@ -10,13 +10,18 @@
 module Data.Morpheus.Server.Types.DirectiveDefinitions
   ( Prefixes (..),
     Deprecated (..),
+    Describe (..),
   )
 where
 
 import Data.Morpheus.Server.Types.Directives (GQLDirective (..))
 import Data.Morpheus.Server.Types.GQLType (GQLType (__type))
 import Data.Morpheus.Server.Types.Internal (mkTypeData)
-import Data.Morpheus.Server.Types.Visitors (VisitEnum, VisitField, VisitType (..))
+import Data.Morpheus.Server.Types.Visitors
+  ( VisitEnum (..),
+    VisitField,
+    VisitType (..),
+  )
 import Data.Morpheus.Types.Internal.AST
   ( DirectiveLocation (..),
   )
@@ -66,3 +71,15 @@ instance GQLDirective Deprecated where
       '[ 'FIELD_DEFINITION,
          'ENUM_VALUE
        ]
+
+newtype Describe = Describe {text :: Text}
+  deriving
+    ( GQLType,
+      Generic
+    )
+
+instance GQLDirective Describe where
+  type DIRECTIVE_LOCATIONS Describe = '[ 'ENUM_VALUE]
+
+instance VisitEnum Describe where
+  visitEnumDescription Describe {text} _ = Just text
